Part 1 - AML monitoring/langchain-workflow/requirements.txt
langchain>=0.0.300
openai>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
pdfminer.six>=20221105
pypdf>=3.0.0
pandas>=1.5.0
pytest>=7.0.0
